Update to xalan 2.7.1.jbossorg-2
xalan:xalan 2.7.1.jbossorg-1 has a broken pom that refers to a non-existent
version of xalan:serializer. xalan:xalan 2.7.1.jbossorg-2 was released
to address that, and should be identical to jbossorg-1 other than that
change (see https://issues.jboss.org/browse/JBBUILD-708).

Updating to jbossorg-2 makes it easier for consumers of the xalan
module.xml to resolve the artifacts it refers to (wildfly-swarm is one
such consumer).