update README.adoc
Updating connection profiles on WildFly, is not yet supported while deployed applications are using the profiles.  This is something that we will work on, by either updating the shared connections while they are in use (hot/live if possible) or adding application dependencies (e.g. similar to JPA pu dependencies or via a new jboss-nosql.xml that specifies NoSQL profile dependency) on the shared connection service.